# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This document defines the RESTful endpoints provided by the GSMA Mobile\
    \ Money API"
  version: "1.0"
  title: "Mobile Money API"
host: "mmapi.gsma.com"
basePath: "/v1.0/MM"
schemes:
- "https"
produces:
- "application/json"
paths:
  /transactions:
    post:
      summary: "Create A Transaction"
      description: "Provided with a valid object representation, this endpoint allows\
        \ for a new transaction to be created"
      operationId: "transactionsPOST"
      consumes:
      - "application/json"
      parameters:
      - name: "X-CorrelationID"
        in: "header"
        description: "Header parameter to uniquely identify the request"
        required: true
        type: "string"
        pattern: "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$"
      - name: "Date"
        in: "header"
        description: "Header parameter to indicate the date and time that the message\
          \ was originated"
        required: true
        type: "string"
        format: "datetime"
      - in: "body"
        name: "requestTransaction"
        description: "Represents the request body of a transaction"
        required: true
        schema:
          $ref: "#/definitions/requestTransaction"
      responses:
        201:
          description: "Represents a Transaction response"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/responseTransaction"
        202:
          description: "Represents the response body of an asynchronous request"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/requestState"
        400:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        401:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        500:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        503:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
      x-swagger-router-controller: "Default"
  /transactions/{transactionReference}:
    get:
      summary: "View A Transaction"
      description: "This endpoint returns the current state of a transaction"
      operationId: "transactionsTransactionReferenceGET"
      parameters:
      - name: "Date"
        in: "header"
        description: "Header parameter to indicate the date and time that the message\
          \ was originated"
        required: true
        type: "string"
        format: "datetime"
      - name: "transactionReference"
        in: "path"
        description: "Path variable to uniquely identify the transaction"
        required: true
        type: "string"
        maxLength: 256
        minLength: 1
      responses:
        200:
          description: "Represents a Transaction response"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/responseTransaction"
        400:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        401:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        500:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        503:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
      x-swagger-router-controller: "Default"
  /transactions/{transactionReference}/reversals:
    post:
      summary: "Create A Reversal"
      description: "Provided with a valid object representation, this endpoint allows\
        \ for a new reversal to be created"
      operationId: "transactionsTransactionReferenceReversalsPOST"
      consumes:
      - "application/json"
      parameters:
      - name: "X-CorrelationID"
        in: "header"
        description: "Header parameter to uniquely identify the request"
        required: true
        type: "string"
        pattern: "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$"
      - name: "Date"
        in: "header"
        description: "Header parameter to indicate the date and time that the message\
          \ was originated"
        required: true
        type: "string"
        format: "datetime"
      - name: "transactionReference"
        in: "path"
        description: "Path variable to uniquely identify the transaction"
        required: true
        type: "string"
        maxLength: 256
        minLength: 1
      - in: "body"
        name: "requestReversal"
        description: "Represents the request body of a reversal"
        required: true
        schema:
          $ref: "#/definitions/requestReversal"
      responses:
        201:
          description: "Represents a Reversal response"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/responseReversal"
        202:
          description: "Represents the response body of an asynchronous request"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/requestState"
        400:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        401:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        500:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
        503:
          description: "error object"
          headers:
            Date:
              type: "string"
              format: "datetime"
              description: "The date and time that the response message was sent"
          schema:
            $ref: "#/definitions/error"
      x-swagger-router-controller: "Default"
...
definitions:
...
  requestTransaction:
    allOf:
    - $ref: "#/definitions/baseTransaction"
    - type: "object"
      properties:
        transactionStatus:
          type: "string"
          description: "Indicates the status of the transaction as stored by the API\
            \ provider."
          minLength: 0
          maxLength: 256
        internationalTransferInformation:
          $ref: "#/definitions/internationalTransfer"
  responseTransaction:
    allOf:
    - $ref: "#/definitions/baseTransaction"
    - type: "object"
      required:
      - "transactionReference"
      - "transactionStatus"
      properties:
        transactionStatus:
          type: "string"
          description: "Indicates the status of the transaction as stored by the API\
            \ provider."
          minLength: 1
          maxLength: 256
        creationDate:
          type: "string"
          format: "date-time"
          description: "Date and time when the transaction was created by the API\
            \ Provider."
        modificationDate:
          type: "string"
          format: "date-time"
          description: "Date and time when the transaction was modified by the API\
            \ Provider."
        transactionReference:
          type: "string"
          description: "Unique reference for the transaction. This is returned in\
            \ the response by API provider."
          minLength: 1
          maxLength: 256
        transactionReceipt:
          type: "string"
          description: "Transaction receipt number as notified to the parties. This\
            \ may differ from the Transaction Reference."
          minLength: 0
          maxLength: 256
        internationalTransferInformation:
          description: "A collection of properties detailed information specifically\
            \ used for international transfers."
          $ref: "#/definitions/internationalTransferResponse"
...
  requestReversal:
    allOf:
    - $ref: "#/definitions/baseReversal"
    - type: "object"
      properties:
        transactionStatus:
          type: "string"
          description: "Indicates the status of the transaction as stored by the API\
            \ provider."
          minLength: 0
          maxLength: 256
...